requirements.txt
click==6.6
decorator==4.0.10
ipython==5.1.0
ipython-genutils==0.1.0
-e git+git@github.com:sumpfgottheit/lanscan.git@7f2d02b4d2771c951a00f21b1bf7ea7eb89ba421#egg=lanscan
netaddr==0.7.18
netifaces==0.10.5
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
requests==2.12.1
scapy-python3==0.18
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
